@@ -37,62 +40,69 @@
 	},
 	"private": true,
 	"dependencies": {
-		"@angular/animations": "19.1.4",
-		"@angular/common": "19.1.4",
-		"@angular/compiler": "19.1.4",
-		"@angular/core": "19.1.4",
-		"@angular/forms": "19.1.4",
-		"@angular/platform-browser": "19.1.4",
-		"@angular/platform-browser-dynamic": "19.1.4",
-		"@angular/router": "19.1.4",
-		"@ngneat/lib": "^3.0.1",
-		"@nx/angular": "20.4.0",
-		"angular-svg-icon": "^16.0.0",
-		"rxjs": "~7.8.0",
+		"@angular/animations": "20.3.9",
+		"@angular/common": "20.3.9",
+		"@angular/compiler": "20.3.9",
+		"@angular/core": "20.3.9",
+		"@angular/forms": "20.3.9",
+		"@angular/platform-browser": "20.3.9",
+		"@angular/platform-browser-dynamic": "20.3.9",
+		"@angular/router": "20.3.9",
+		"@ngneat/lib": "^5.0.0",
+		"@nx/angular": "22.1.3",
+		"angular-svg-icon": "^19.1.1",
+		"rxjs": "~7.8.2",
 		"tslib": "^2.3.0",
 		"zone.js": "0.15.0"
 	},
 	"devDependencies": {
-		"@angular-devkit/build-angular": "19.1.5",
-		"@angular-devkit/core": "19.1.5",
-		"@angular-devkit/schematics": "19.1.5",
-		"@angular-eslint/eslint-plugin": "17.3.0",
-		"@angular-eslint/eslint-plugin-template": "17.3.0",
-		"@angular-eslint/template-parser": "17.3.0",
-		"@angular/cli": "~19.1.0",
-		"@angular/compiler-cli": "19.1.4",
-		"@angular/language-service": "19.1.4",
-		"@nx/cypress": "20.4.0",
-		"@nx/eslint": "20.4.0",
-		"@nx/eslint-plugin": "20.4.0",
-		"@nx/jest": "20.4.0",
-		"@nx/js": "20.4.0",
-		"@nx/web": "20.4.0",
-		"@nx/workspace": "20.4.0",
-		"@schematics/angular": "19.1.5",
+		"@angular-devkit/build-angular": "20.3.13",
+		"@angular-devkit/core": "20.3.13",
+		"@angular-devkit/schematics": "20.3.13",
+		"@angular-eslint/builder": "^20.7.0",
+		"@angular-eslint/eslint-plugin": "20.7.0",
+		"@angular-eslint/eslint-plugin-template": "20.7.0",
+		"@angular-eslint/template-parser": "20.7.0",
+		"@angular/cli": "~20.3.13",
+		"@angular/compiler-cli": "20.3.9",
+		"@angular/language-service": "20.3.9",
+		"@nx/cypress": "22.1.3",
+		"@nx/eslint": "22.1.3",
+		"@nx/eslint-plugin": "22.1.3",
+		"@nx/jest": "22.1.3",
+		"@nx/js": "22.1.3",
+		"@nx/web": "22.1.3",
+		"@nx/workspace": "22.1.3",
+		"@schematics/angular": "20.3.9",
 		"@types/jest": "29.5.14",
-		"@types/node": "18.16.9",
-		"@typescript-eslint/eslint-plugin": "7.18.0",
-		"@typescript-eslint/parser": "7.18.0",
-		"autoprefixer": "^10.4.0",
-		"cypress": "13.17.0",
-		"eslint": "8.46.0",
-		"eslint-config-prettier": "9.0.0",
-		"eslint-plugin-cypress": "2.15.2",
+		"@types/node": "22.19.1",
+		"@typescript-eslint/eslint-plugin": "8.48.1",
+		"@typescript-eslint/parser": "8.48.1",
+		"autoprefixer": "^10.4.22",
+		"cypress": "15.7.1",
+		"eslint": "9.39.1",
+		"eslint-config-prettier": "10.1.8",
+		"eslint-plugin-cypress": "5.2.0",
 		"git-cz": "^3.2.1",
 		"jest": "29.7.0",
 		"jest-environment-jsdom": "29.7.0",
-		"jest-preset-angular": "14.4.2",
-		"ng-packagr": "19.1.2",
-		"nx": "20.4.0",
+		"jest-preset-angular": "14.6.2",
+		"karma": "^6.4.4",
+		"karma-chrome-launcher": "~3.2.0",
+		"karma-coverage": "~2.2.1",
+		"karma-coverage-istanbul-reporter": "^3.0.3",
+		"karma-jasmine": "~5.1.0",
+		"karma-jasmine-html-reporter": "^2.1.0",
+		"ng-packagr": "20.3.2",
+		"nx": "22.1.3",
 		"postcss": "^8.3.9",
 		"postcss-import": "14.1.0",
 		"postcss-preset-env": "7.5.0",
 		"postcss-url": "10.1.3",
-		"prettier": "2.7.1",
+		"prettier": "3.7.4",
 		"standard-version": "^9.3.2",
-		"ts-jest": "29.1.0",
-		"ts-node": "10.9.1",
-		"typescript": "5.7.3"
+		"ts-jest": "29.4.6",
+		"ts-node": "10.9.2",
+		"typescript": "5.9.3"
 	}
 }
