@@ -1,5 +1,5 @@
 import { HttpClient } from '@angular/common/http';
-import { Injectable, Optional } from '@angular/core';
+import { Injectable, inject } from '@angular/core';
 import { SvgIconRegistryService } from 'angular-svg-icon';
 import { EMPTY, Observable, Subject } from 'rxjs';
 import { catchError, take, tap } from 'rxjs/operators';
@@ -10,6 +10,9 @@ import { CustomIconData, IconImageFile } from '../icon.interface';
 	providedIn: 'root',
 })
 export class AngularSvgIconPreloaderService {
+	private _http = inject(HttpClient);
+	private _iconRegistry = inject(SvgIconRegistryService);
+
 	private configUrl: string = './assets/icons.json';
 	private iconsFileData: {
 		iconImageFiles: IconImageFile[];
@@ -17,11 +20,11 @@ export class AngularSvgIconPreloaderService {
 	} = { customIcons: [], iconImageFiles: [] };
 	public configSubject: Subject<any> = new Subject<any>();
 
-	constructor(
-		private _http: HttpClient,
-		@Optional() config: AngularSvgIconPreloaderConfig,
-		private _iconRegistry: SvgIconRegistryService
-	) {
+	constructor() {
+		const config = inject(AngularSvgIconPreloaderConfig, {
+			optional: true,
+		});
+
 		if (config && config.configUrl) {
 			this.configUrl = config.configUrl;
 		}
@@ -41,13 +44,13 @@ export class AngularSvgIconPreloaderService {
 					}) => {
 						this.iconsFileData = configData;
 						this.loadIcons();
-					}
+					},
 				),
 				catchError((err) => {
 					console.error(
 						'An error occurred loading the icons:\n',
 						err,
-						'\nNo icons will be loaded.'
+						'\nNo icons will be loaded.',
 					);
 					this.iconsFileData = {
 						customIcons: [],
@@ -55,7 +58,7 @@ export class AngularSvgIconPreloaderService {
 					};
 					this.loadIcons();
 					return EMPTY;
-				})
+				}),
 			);
 	}
 
