@@ -1,11 +1,11 @@
 import { Component } from '@angular/core';
 
 @Component({
-    selector: 'angular-svg-icon-preloader-root',
-    templateUrl: './app.component.html',
-    styleUrls: ['./app.component.scss'],
-    standalone: false
+	selector: 'angular-svg-icon-preloader-root',
+	templateUrl: './app.component.html',
+	styleUrls: ['./app.component.scss'],
+	standalone: false,
 })
 export class AppComponent {
-  title = 'angular-svg-icon-preloader-demo';
+	title = 'angular-svg-icon-preloader-demo';
 }
