@@ -27,12 +27,12 @@ const cp = require('child_process');
 const isWindows = os.platform() === 'win32';
 let output;
 try {
-  output = require('@nrwl/workspace').output;
+	output = require('@nrwl/workspace').output;
 } catch (e) {
-  console.warn(
-    'Angular CLI could not be decorated to enable computation caching. Please ensure @nrwl/workspace is installed.'
-  );
-  process.exit(0);
+	console.warn(
+		'Angular CLI could not be decorated to enable computation caching. Please ensure @nrwl/workspace is installed.',
+	);
+	process.exit(0);
 }
 
 /**
@@ -44,20 +44,20 @@ const angularCLIInitPath = 'node_modules/@angular/cli/lib/cli/index.js';
  * Patch index.js to warn you if you invoke the undecorated Angular CLI.
  */
 function patchAngularCLI(initPath) {
-  const angularCLIInit = fs.readFileSync(initPath, 'utf-8').toString();
+	const angularCLIInit = fs.readFileSync(initPath, 'utf-8').toString();
 
-  if (!angularCLIInit.includes('NX_CLI_SET')) {
-    fs.writeFileSync(
-      initPath,
-      `
+	if (!angularCLIInit.includes('NX_CLI_SET')) {
+		fs.writeFileSync(
+			initPath,
+			`
 if (!process.env['NX_CLI_SET']) {
   const { output } = require('@nrwl/workspace');
   output.warn({ title: 'The Angular CLI was invoked instead of the Nx CLI. Use "npx ng [command]" or "nx [command]" instead.' });
 }
 ${angularCLIInit}
-    `
-    );
-  }
+    `,
+		);
+	}
 }
 
 /**
@@ -65,40 +65,43 @@ ${angularCLIInit}
  * invoke the Nx CLI and get the benefits of computation caching.
  */
 function symlinkNgCLItoNxCLI() {
-  try {
-    const ngPath = './node_modules/.bin/ng';
-    const nxPath = './node_modules/.bin/nx';
-    if (isWindows) {
-      /**
-       * This is the most reliable way to create symlink-like behavior on Windows.
-       * Such that it works in all shells and works with npx.
-       */
-      ['', '.cmd', '.ps1'].forEach((ext) => {
-        if (fs.existsSync(nxPath + ext))
-          fs.writeFileSync(ngPath + ext, fs.readFileSync(nxPath + ext));
-      });
-    } else {
-      // If unix-based, symlink
-      cp.execSync(`ln -sf ./nx ${ngPath}`);
-    }
-  } catch (e) {
-    output.error({
-      title:
-        'Unable to create a symlink from the Angular CLI to the Nx CLI:' +
-        e.message,
-    });
-    throw e;
-  }
+	try {
+		const ngPath = './node_modules/.bin/ng';
+		const nxPath = './node_modules/.bin/nx';
+		if (isWindows) {
+			/**
+			 * This is the most reliable way to create symlink-like behavior on Windows.
+			 * Such that it works in all shells and works with npx.
+			 */
+			['', '.cmd', '.ps1'].forEach((ext) => {
+				if (fs.existsSync(nxPath + ext))
+					fs.writeFileSync(
+						ngPath + ext,
+						fs.readFileSync(nxPath + ext),
+					);
+			});
+		} else {
+			// If unix-based, symlink
+			cp.execSync(`ln -sf ./nx ${ngPath}`);
+		}
+	} catch (e) {
+		output.error({
+			title:
+				'Unable to create a symlink from the Angular CLI to the Nx CLI:' +
+				e.message,
+		});
+		throw e;
+	}
 }
 
 try {
-  symlinkNgCLItoNxCLI();
-  patchAngularCLI(angularCLIInitPath);
-  output.log({
-    title: 'Angular CLI has been decorated to enable computation caching.',
-  });
+	symlinkNgCLItoNxCLI();
+	patchAngularCLI(angularCLIInitPath);
+	output.log({
+		title: 'Angular CLI has been decorated to enable computation caching.',
+	});
 } catch (e) {
-  output.error({
-    title: 'Decoration of the Angular CLI did not complete successfully',
-  });
+	output.error({
+		title: 'Decoration of the Angular CLI did not complete successfully',
+	});
 }
