@@ -1,3 +1,3 @@
-const nxPreset = require('@nrwl/jest/preset');
+const nxPreset = require('@nx/jest/preset');
 
 module.exports = { ...nxPreset };
