@@ -1,9 +1,9 @@
 export interface IconImageFile {
-  iconName: string;
-  iconPath: string;
+	iconName: string;
+	iconPath: string;
 }
 
 export interface CustomIconData {
-  iconName: string;
-  iconData: string;
+	iconName: string;
+	iconData: string;
 }
