@@ -1,6 +1,6 @@
 module.exports = {
-  singleQuote: true,
-  tabWidth: 2,
-  useTabs: false,
-  printWidth: 120,
+	singleQuote: true,
+	tabWidth: 2,
+	useTabs: false,
+	printWidth: 120,
 };
